% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{ancestors}
\alias{ancestors}
\alias{ancestors.LatentDigraphFixedOrder}
\alias{ancestors.LatentDigraph}
\alias{ancestors.MixedGraph}
\title{All ancestors of a collection of nodes}
\usage{
ancestors(this, nodes, ...)

\method{ancestors}{LatentDigraphFixedOrder}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{ancestors}{LatentDigraph}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{ancestors}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object}

\item{nodes}{the nodes from which to find all ancestors}

\item{...}{ignored.}

\item{includeObserved}{if TRUE includes observed nodes in the returned set.}

\item{includeLatents}{if TRUE includes latent nodes in the returned set.}
}
\value{
the ancestors of the nodes in the observed part of the graph.
}
\description{
Finds all the ancestors of a collection of nodes. These ancestors DO
include the nodes themselves (every node is considered an ancestor of itself).
}
