% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalID.R
\name{bidirectedComponents}
\alias{bidirectedComponents}
\title{Get bidirected components of a mixed graph}
\usage{
bidirectedComponents(graph)
}
\arguments{
\item{graph}{a \code{\link{MixedGraph}} object representing
the mixed graph.}
}
\value{
list, where each object is a MixedGraph with at least two nodes.
}
\description{
Returns induced subgraphs of connected bidirected components with more than 1 node.
}
