% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_rules.R
\name{cfa_rules}
\alias{cfa_rules}
\alias{rule_cfa_three_indicator}
\alias{rule_cfa_two_indicator}
\title{Rules for confirmatory factor analysis models}
\usage{
rule_cfa_three_indicator(partable)

rule_cfa_two_indicator(partable)
}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Rules for confirmatory factor analysis models
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).

Kenny, D. A. (1979). Correlation and causality.
}
\keyword{internal}
