% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{check_recursion}
\alias{check_recursion}
\title{Check recursion using Depth-First Search algorithm}
\usage{
check_recursion(partable, x)
}
\arguments{
\item{partable}{A lavaan parameter table}

\item{x}{A character vector of variable names from which to start the algorithm.}
}
\description{
Check recursion using Depth-First Search algorithm
}
\keyword{internal}
