% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_localBBCriterion.R
\name{checkLocalBBCriterion}
\alias{checkLocalBBCriterion}
\title{Check Local BB-Criterion}
\usage{
checkLocalBBCriterion(adjMatrix, latentNodes, observedNodes)
}
\arguments{
\item{adjMatrix}{of the graph}

\item{latentNodes}{set of latent nodes}

\item{observedNodes}{set of observed nodes}
}
\value{
a list consisting of a Boolean, whether the graph is
        sign-identifiable and if yes, a list consisting of the sets
}
\description{
for a set of latent nodes, find a set of observed nodes U so that the tuple
satisfies the local BB-criterion and the inequality concerning the
cardinalities of the sets
}
\references{
Sturma, N., Kranzlmüller, M., Portakal, I., and Drton, M.  (2025) Matching
Criterion for Identifiability in Sparse Factor Analysis.
arXiv:2502.02986
}
