% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{children}
\alias{children}
\alias{children.LatentDigraphFixedOrder}
\alias{children.LatentDigraph}
\alias{children.MixedGraph}
\title{All children of a collection of nodes.}
\usage{
children(this, nodes, ...)

\method{children}{LatentDigraphFixedOrder}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{children}{LatentDigraph}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{children}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object.}

\item{nodes}{nodes the nodes of which to find the children}

\item{...}{ignored.}

\item{includeObserved}{if TRUE includes observed nodes in the returned set.}

\item{includeLatents}{if TRUE includes latent nodes in the returned set.}
}
\value{
the observed children
}
\description{
Returns all children of the collection (does not necessarily include the
input nodes themselves unless they are parents of one another).
}
