% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_model.R
\name{classify_model}
\alias{classify_model}
\title{Classify a model using the parameter table for internal use}
\usage{
classify_model(partable = NULL)
}
\description{
Classify a model using the parameter table for internal use
}
\author{
Zach Vig
}
\keyword{internal}
