% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgewiseID.R
\name{createEdgewiseIdentifier}
\alias{createEdgewiseIdentifier}
\title{Create an edgewise identification function}
\usage{
createEdgewiseIdentifier(idFunc, sources, targets, node, solvedNodeParents,
  sourceParentsToRemove)
}
\value{
an identification function
}
\description{
A helper function for \link{\code{edgewiseIdentifyStep}}, creates an
identifier function based on its given parameters. This created identifier
function will identify the directed edges from 'targets' to 'node.'
}

