% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{createHalfTrekFlowGraph}
\alias{createHalfTrekFlowGraph}
\alias{createHalfTrekFlowGraph.MixedGraphFixedOrder}
\title{Helper function to create a flow graph.}
\usage{
createHalfTrekFlowGraph(this)

\method{createHalfTrekFlowGraph}{MixedGraphFixedOrder}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Helper function to create a flow graph.
}
