% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{createHtrGraph}
\alias{createHtrGraph}
\alias{createHtrGraph.MixedGraphFixedOrder}
\title{Helper function to create a graph encoding htr relationships.}
\usage{
createHtrGraph(this)

\method{createHtrGraph}{MixedGraphFixedOrder}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Helper function to create a graph encoding htr relationships.
}
