% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSimpleBiDirIdentifier}
\alias{createSimpleBiDirIdentifier}
\title{Identify bidirected edges if all directed edges are identified}
\usage{
createSimpleBiDirIdentifier(idFunc)
}
\arguments{
\item{idFunc}{an identifier function that identifies all directed edges}
}
\value{
a new identifier function that identifies everything.
}
\description{
Creates an identifier function that assumes that all directed edges have
already been identified and then is able to identify all bidirected edges
simultaneously.
}
