% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{createTrekFlowGraph}
\alias{createTrekFlowGraph}
\alias{createTrekFlowGraph.LatentDigraphFixedOrder}
\title{Helper function to create a flow graph.}
\usage{
createTrekFlowGraph(this, ...)

\method{createTrekFlowGraph}{LatentDigraphFixedOrder}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Helper function to create a flow graph.
}
