% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekSepId.R
\name{createTrekSeparationIdentifier}
\alias{createTrekSeparationIdentifier}
\title{Create an trek separation identification function}
\usage{
createTrekSeparationIdentifier(idFunc, sources, targets, node, parent,
  solvedParents)
}
\value{
an identification function
}
\description{
A helper function for \link{\code{trekSeparationIdentifyStep}}, creates an
identifier function based on its given parameters. This created identifier
function will identify the directed edge from 'parent' to 'node.'
}

