% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{createTrGraph}
\alias{createTrGraph}
\alias{createTrGraph.MixedGraphFixedOrder}
\title{Helper function to create a graph encoding trek reachable relationships.}
\usage{
createTrGraph(this)

\method{createTrGraph}{MixedGraphFixedOrder}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Helper function to create a graph encoding trek reachable relationships.
}
