% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{descendants}
\alias{descendants}
\alias{descendants.MixedGraphFixedOrder}
\alias{descendants.MixedGraph}
\title{Get descendants of a node}
\usage{
descendants(this, node)

\method{descendants}{MixedGraphFixedOrder}(this, node)

\method{descendants}{MixedGraph}(this, node)
}
\arguments{
\item{this}{the mixed graph object}

\item{node}{the node from which to get the descendants.}
}
\description{
Finds all descendants of a node, this DOES include the node itself (every
node is considered a descendant of itself).
}
