% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_rules.R
\name{exogenous_x_rule}
\alias{exogenous_x_rule}
\title{Exogenous X rule/MIMIC rules}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Exogenous X rule/MIMIC rules
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen & Davis (2009). Two Rules of Identification for Structural Equation Models.
}
\author{
Zach Vig
}
\keyword{internal}
