% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_ZUTA.R
\name{findColumnsWithSumOne}
\alias{findColumnsWithSumOne}
\title{A Helper Function for Check ZUTA}
\usage{
findColumnsWithSumOne(cleanMatrix)
}
\arguments{
\item{cleanMatrix}{a matrix without zero columns and rows}
}
\value{
a Boolean whether it finds a column with sum=1 and the new matrix,
         without the row in which the entry is, fulfills ZUTA
}
\description{
iterative function that tries to find a column with sum=1, delete the row
with the 1 in that column, check smaller matrix
}
