% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowGraph.R
\name{flowBetween}
\alias{flowBetween}
\alias{flowBetween.FlowGraph}
\title{Flow from one set of nodes to another.}
\usage{
flowBetween(this, sources, sinks)

\method{flowBetween}{FlowGraph}(this, sources, sinks)
}
\arguments{
\item{this}{the flow graph object}

\item{sources}{the nodes from which flow should start.}

\item{sinks}{the nodes at which the flow should end.}
}
\value{
a list with two named components, \code{value} (the size of the
        computed flow) and \code{activeSources} (a vector representing the
        subset of sources which have non-zero flow out of them for the found
        max-flow).
}
\description{
Flow from one set of nodes to another.
}
