% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowGraph.R
\name{FlowGraph}
\alias{FlowGraph}
\title{Construct FlowGraph object}
\usage{
FlowGraph(L = matrix(0,1,1), vertexCaps = 1, edgeCaps = matrix(1,1,1))
}
\arguments{
\item{L}{the adjacency matrix for the flow graph. The (i,j)th of L should be
a 1 if there is an edge from i to j and 0 otherwise.}

\item{vertexCaps}{the capacity of the vertices in the flow graph, should
either be a single number or a vector whose ith entry is the capacity
of vertex i.}

\item{edgeCaps}{the capacities of the edges in the the flow graph, should
be a matrix of the same dimensions as L with (i,j)th entry the
capacity of the i->j edge.}
}
\value{
An object representing the FlowGraph.
}
\description{
Creates an object representing a flow graph.
}
