% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{fully_recursive_rule}
\alias{fully_recursive_rule}
\title{Fully Recursive model rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Fully Recursive model rule
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
