% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generalGenericID}
\alias{generalGenericID}
\title{A general generic identification algorithm template.}
\usage{
generalGenericID(L, O, idStepFunctions, tianDecompose = T)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1 and
the lack of an edge between i and j is encoded as L[i,j]=0. There should be
no directed self loops, i.e. no i such that L[i,i]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path diagram/mixed
graph. Edges are encoded as for the L parameter. Again there should be no
self loops. Also this matrix will be coerced to be symmetric so it is only
necessary to specify an edge once, i.e. if O[i,j]=1 you may, but are not
required to, also have O[j,i]=1.}

\item{idStepFunctions}{a list of identification step functions}

\item{tianDecompose}{True if the mixed graph be decomposed into Tian's
C-components before running the identification algorithms.}
}
\value{
see the return of \code{\link{htcID}}.
}
\description{
A function that encapsulates the general structure of our algorithms for
testing generic identifiability. Allows for various identification algorithms
to be used in concert, in particular it will use the identifier functions
in the list \code{idStepFunctions} sequentially until it can find no more
identifications. The step functions that are currently available for use are
in \code{idStepFunctions}
\enumerate{
  \item htcIdentifyStep
  \item ancestralIdentifyStep
  \item edgewiseIdentifyStep
  \item trekSeparationIdentifyStep
}
}

