% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{get_rules}
\alias{get_rules}
\title{Gather identification rules as a list}
\usage{
get_rules(rule = "*", model_type = "*")
}
\arguments{
\item{rule}{A string specifying the name of the rule as it's defined
in the package. Use "*" to get all rules (or all rules of the defined
model type)}

\item{model_type}{A string specifying the model type from which to get
rules. Types include "reg" (simultaneous equation models/regression models),
"cfa" (confirmatory factor analysis models), or "sem" (general SEM models).
Use "*" to get from all model types}
}
\value{
A list object with the rule function(s) specified by the user.
}
\description{
The `semID` package stores identification rule functions internally. This function
 makes them available to the user as a list.
}
\author{
Zach Vig
}
