% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestralID.R
\name{getAncestors}
\alias{getAncestors}
\title{Get getAncestors of nodes in a graph.}
\usage{
getAncestors(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the getAncestors.}
}
\value{
a sorted vector of all ancestor nodes.
}
\description{
Get the getAncestors of a collection of nodes in a graph g, the getAncestors DO
include the the nodes themselves.
}
