% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{getMixedGraph}
\alias{getMixedGraph}
\alias{getMixedGraph.LatentDigraph}
\title{Get the corresponding mixed graph}
\usage{
getMixedGraph(this, ...)

\method{getMixedGraph}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the graph object}
}
\description{
Get the corresponding mixed graph
}
