% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestralID.R
\name{getParents}
\alias{getParents}
\title{Get getParents of nodes in a graph.}
\usage{
getParents(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the getParents.}
}
\value{
a sorted vector of all parent nodes.
}
\description{
Get the getParents of a collection of nodes in a graph g, the getParents DO include
the input nodes themselves.
}
