% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestralID.R
\name{getSiblings}
\alias{getSiblings}
\title{Get getSiblings of nodes in a graph.}
\usage{
getSiblings(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the getSiblings.}
}
\value{
a sorted vector of all getSiblings of nodes.
}
\description{
Get the getSiblings of a collection of nodes in a graph g, the getSiblings DO
include the input nodes themselves.
}
