% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{graphID.ancestral}
\alias{graphID.ancestral}
\title{Determine generic identifiability of an acyclic mixed graph using ancestral
decomposition.}
\usage{
graphID.ancestral(L, O)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path
diagram/mixed graph.}
}
\value{
The vector of nodes that could be determined to be generically
identifiable using the above algorithm.
}
\description{
For an input, acyclic, mixed graph attempts to determine if the graph is
generically identifiable using decomposition by ancestral subsets. See
algorithm 1 of Drton and Weihs (2015).
}
\references{
{Drton}, M. and {Weihs}, L. (2015) Generic Identifiability of Linear
Structural Equation Models by Ancestor Decomposition. arXiv 1504.02992
}

