% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestralID.R
\name{graphID.ancestralID}
\alias{graphID.ancestralID}
\title{Determine generic identifiability of an acyclic mixed graph using ancestral
decomposition.}
\usage{
graphID.ancestralID(L, O)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1 and
the lack of an edge between i and j is encoded as L[i,j]=0. There should be
no directed self loops, i.e. no i such that L[i,i]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path diagram/mixed
graph. Edges are encoded as for the L parameter. Again there should be no
self loops. Also this matrix will be coerced to be symmetric so it is only
necessary to specify an edge once, i.e. if O[i,j]=1 you may, but are not
required to, also have O[j,i]=1.}
}
\value{
The vector of nodes that could be determined to be generically
identifiable using the above algorithm.
}
\description{
For an input, acyclic, mixed graph attempts to determine if the graph is
generically identifiable using decomposition by ancestral subsets. See
algorithm 1 of Drton and Weihs (2015).
}
\references{
{Drton}, M. and {Weihs}, L. (2015) Generic Identifiability of Linear
Structural Equation Models by Ancestor Decomposition. arXiv 1504.02992
}
