% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{graphID.HTC}
\alias{graphID.HTC}
\title{Determines if a mixed graph is HTC-identifiable.}
\usage{
graphID.HTC(L, O)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path
diagram/mixed graph.}
}
\value{
The vector of HTC-identifiable nodes.
}
\description{
Uses the half-trek criterion of Foygel, Draisma, and Drton (2015) to check
if an input mixed graph is generically identifiable.
}
\references{
Foygel, R., Draisma, J., and Drton, M.  (2012) Half-trek criterion for
generic identifiability of linear structural equation models.
\emph{Ann. Statist.} 40(3): 1682-1713.
}

