% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{graphID.nonID}
\alias{graphID.nonID}
\title{Check for generic infinite-to-one via the half-trek criterion.}
\usage{
graphID.nonID(L, O)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path
diagram/mixed graph.}
}
\description{
Checks if a mixed graph is infinite-to-one using the half-trek criterion
presented by Foygel, Draisma, and Drton (2012).
}
\references{
Foygel, R., Draisma, J., and Drton, M.  (2012) Half-trek criterion for
generic identifiability of linear structural equation models.
\emph{Ann. Statist.} 40(3): 1682-1713.
}

