% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htcID.R
\name{htcID}
\alias{htcID}
\title{Determines which edges in a mixed graph are HTC-identifiable.}
\usage{
htcID(L, O, tianDecompose = T)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1 and
the lack of an edge between i and j is encoded as L[i,j]=0. There should be
no directed self loops, i.e. no i such that L[i,i]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path diagram/mixed
graph. Edges are encoded as for the L parameter. Again there should be no
self loops. Also this matrix will be coerced to be symmetric so it is only
necessary to specify an edge once, i.e. if O[i,j]=1 you may, but are not
required to, also have O[j,i]=1.}

\item{tianDecompose}{TRUE or FALSE determining whether or not the Tian
decomposition should be used before running the
current generic identification algorithm. In general
letting this be TRUE will make the algorithm faster and
more powerful.}
}
\value{
a list with 5 names arguments:
\describe{
  \item{\code{solvedParents}}{a list whose ith element contains a vector
  containing the subsets of parents of node i for which the edge j->i could
  be shown to be generically identifiable.}
  \item{\code{unsolvedParents}}{as for \code{solvedParents} but for the
  unsolved parents.}
  \item{\code{solvedSiblings}}{as for \code{solvedParents} but for the
  siblings of node i (i.e. the bidirected neighbors of i).}
  \item{\code{unsolvedSiblings}}{as for \code{solvedSilbings} but for the
  unsolved siblings of node i (i.e. the bidirected neighbors of i).}
  \item{\code{identifier}}{a function that takes a (generic) covariance
  matrix corresponding to the graph and identifies the edges parameters
  from solvedParents and solvedSiblings. See \code{\link{htcIdentifyStep}}
  for a more in-depth discussion of identifier functions.}
}
}
\description{
Uses the half-trek criterion of Foygel, Draisma, and Drton (2013) determine
which edges in a mixed graph are generically identifiable. Depending on your
application it faster to use the \code{\link{graphID.htcID}} function
instead of this one, this function has the advantage of returning additional
information.
}
\references{
Foygel, R., Draisma, J., and Drton, M.  (2012) Half-trek criterion for
generic identifiability of linear structural equation models.
\emph{Ann. Statist.} 40(3): 1682-1713.

Jin Tian. 2005. Identifying direct causal effects in linear models. In
\emph{Proceedings of the 20th national conference on Artificial intelligence
- Volume 1} (AAAI'05), Anthony Cohn (Ed.), Vol. 1. AAAI Press 346-352.
}

