% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{htr}
\alias{htr}
\title{Get all HTR nodes from a set of nodes in a graph.}
\usage{
htr(dG, bG, nodes)
}
\arguments{
\item{dG}{a directed graph representing the directed part of the mixed graph.}

\item{bG}{an undirected graph representing the undirected part of the mixed
graph.}

\item{nodes}{the nodes in the graph of which to get the HTR nodes.}
}
\value{
a sorted list of all half-trek reachable nodes.
}
\description{
Gets all vertices in a graph that are half-trek reachable from a set of
nodes.
WARNING: Often the half-trek reachable nodes from a vertex v are defined to
not include the vertex v or its siblings. We DO NOT follow this convention,
the returned set will include input nodes and their siblings.
}

