% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{htrFrom}
\alias{htrFrom}
\alias{htrFrom.MixedGraphFixedOrder}
\alias{htrFrom.MixedGraph}
\title{Half trek reachable nodes.}
\usage{
htrFrom(this, node)

\method{htrFrom}{MixedGraphFixedOrder}(this, node)

\method{htrFrom}{MixedGraph}(this, node)
}
\arguments{
\item{this}{the mixed graph object}

\item{node}{the node from which to get all half-trek reachable nodes.}
}
\value{
a vector of all nodes half-trek reachable from node.
}
\description{
Half trek reachable nodes.
}
