% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\alias{id.semscale}
\alias{id.lavaan}
\alias{id.character}
\alias{id.data.frame}
\alias{id.default}
\alias{id2}
\title{Evaluate common Structural Equation Model (SEM) identification rules}
\usage{
id(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)

\method{id}{semscale}(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)

\method{id}{lavaan}(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)

\method{id}{character}(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)

\method{id}{data.frame}(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)

\method{id}{default}(x, include.msgs = TRUE, call = "sem", ...)

id2(x, include.msgs = TRUE, call = "sem", ...)
}
\arguments{
\item{x}{A character string model in \code{lavaan} syntax, a
\code{lavaan} parameter table, or a fitted \code{lavaan} object.}

\item{include.msgs}{Logical. If \code{TRUE}, the output will include why a rule 
does not pass or is not applicable, along with any other helpful information.
Default: \code{TRUE}.}

\item{call}{A character string specifying the call you intend to use to fit
the model. This will ensure the correct model defaults are specified. Options
currently include "lavaan", "sem", or "cfa". If a parameter table for fit
object are supplied, this argument is ignored. Default: "sem".}

\item{twostep}{A logical indicating whether to use the two-step identification rule
instead of the usual one-step. See details. Default: \code{FALSE}.}

\item{...}{Additional arguments passed to the \code{lavaanify} function from
\code{lavaan}. See \link[lavaan]{lavaanify} for more information. If parameter
tables or fitted model objects are supplied, these arguments are ignored.}
}
\value{
An object of class \code{semid} or \code{semid2} (if \code{twostep = TRUE}
or \code{id2} is called. See details.)
}
\description{
This is the "workhorse" function of the \code{semID} package. The user supplies
a model string in \code{lavaan} syntax (see \link[lavaan]{model.syntax} for more
details) and the function prints an informative table to the console about the
status of the model on a variety of common identification rules.
}
\details{
The primary output of \code{id()} is a table printed to the console, where rows
correspond to rules, and columns include "Pass" (did the rule pass?), "Necessary"
(is the rule necessary for identification?), and "Sufficient" (is the rule
sufficient for identification?). These columns can take values "Yes", "No",
or be left blank in the case of NA values.

If the user set the \code{warn} argument to "TRUE" (which is the default),
a column labeled "Warning" is appended to the table, and an output section called
"Warnings" is printed below the table. Warnings include why a rule failed, why
a rule is not applicable to the current model, or why the necessary/sufficient
conditions may not apply as usual.

Warnings are identified by a number, and corresponding warning numbers are listed
in the "Warning" column of the table.

\code{call} takes character values "lavaan", "sem", or "cfa", specifying which
\code{lavaan} function the user intends to call (and thus which defaults should)
be used) when fitting the model in the case a model string is supplied. Supplying
a parameter table or fitted model object ignores the \code{call} argument since
defaults will have already been implemented.

\code{id2} is a wrapper function for calling \code{id} with argument \code{twostep}
set to \code{TRUE}. The two-step method parses an SEM into a CFA model and a latent
variable/structural model, and evaluates the identification rules on each. If both
parts are identified, the whole model is identified.
}
\examples{
# Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
id(HS.model, include.msgs = TRUE, call = "cfa", 
   meanstructure = FALSE)

}
