% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id_mplus}
\alias{id_mplus}
\title{Evaluate identification rules for an Mplus model}
\usage{
id_mplus(x, include.msgs = TRUE, call = "sem", twostep = FALSE, ...)
}
\arguments{
\item{x}{A character string model in Mplus syntax, or a path to an Mplus input file.}

\item{include.msgs}{Logical. If \code{TRUE}, the output will include why a rule 
does not pass or is not applicable, along with any other helpful information.
Default: \code{TRUE}.}

\item{call}{A character string specifying the call you intend to use to fit
the model. This will ensure the correct model defaults are specified. Options
currently include "lavaan", "sem", or "cfa". If a parameter table for fit
object are supplied, this argument is ignored. Default: "sem".}

\item{twostep}{A logical indicating whether to use the two-step identification rule
instead of the usual one-step. See details. Default: \code{FALSE}.}

\item{...}{Additional arguments passed to the \code{lavaanify} function from
\code{lavaan}. See \link[lavaan]{lavaanify} for more information. If parameter
tables or fitted model objects are supplied, these arguments are ignored.}
}
\description{
This is a wrapper function for \code{\link{id}} that accomodates Mplus model syntax
(as a string) or Mplus input files (with extension ".inp").
}
