% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{inducedSubgraph}
\alias{inducedSubgraph}
\alias{inducedSubgraph.LatentDigraph}
\alias{inducedSubgraph.MixedGraph}
\title{Get the induced subgraph on a collection of nodes}
\usage{
inducedSubgraph(this, nodes, ...)

\method{inducedSubgraph}{LatentDigraph}(this, nodes, ...)

\method{inducedSubgraph}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object}

\item{nodes}{the nodes on which to create the induced subgraph.}

\item{...}{ignored.}
}
\description{
Get the induced subgraph on a collection of nodes
}
