% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{isSibling}
\alias{isSibling}
\alias{isSibling.MixedGraph}
\title{Are two nodes siblings?}
\usage{
isSibling(this, node1, node2, ...)

\method{isSibling}{MixedGraph}(this, node1, node2, ...)
}
\arguments{
\item{this}{the mixed graph object}

\item{node1}{a node}

\item{node2}{a second node}

\item{...}{ignored.}
}
\value{
TRUE if the nodes are siblings in the graph, FALSE otherwise
}
\description{
Are two nodes siblings?
}
