% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{L}
\alias{L}
\alias{L.MixedGraph}
\title{Get adjacency matrix for directed part.}
\usage{
L(this)

\method{L}{MixedGraph}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Get adjacency matrix for directed part.
}
