% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_rules.R
\name{latent_scaling_rule}
\alias{latent_scaling_rule}
\title{Latent Scaling rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Latent Scaling rule
}
\references{
Bollen, K. A., Lilly, A. G., & Luo, L. (2024). Selecting scaling
indicators in structural equation models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
