% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{LatentDigraphFixedOrder}
\alias{LatentDigraphFixedOrder}
\title{Construct LatentDigraphFixedOrder object}
\usage{
LatentDigraphFixedOrder(L = matrix(0,1,1), numObserved = nrow(L))
}
\arguments{
\item{L}{see \code{\link{graphID}} for the appropriate form of L. The
first numObserved rows of L correspond to the observed nodes in the
graph, all other nodes are considered unobserved.}

\item{numObserved}{a non-negative integer representing the number of observed
nodes in the graph.}
}
\value{
An object representing the LatentDigraphFixedOrder
}
\description{
Creates an object representing a directed graph with some number of
nodes which are latent (unobserved).
}
