% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfhtcID.R
\name{latentDigraphHasSimpleNumbering}
\alias{latentDigraphHasSimpleNumbering}
\title{Checks that a LatentDigraph has appropriate node numbering}
\usage{
latentDigraphHasSimpleNumbering(graph)
}
\arguments{
\item{graph}{a \code{\link{LatentDigraph}} object representing
the latent-factor graph. All latent nodes in this graph should be
source nodes (i.e. have no parents).}
}
\description{
Checks that the input latent digraph has nodes numbered from 1
to latentDigraph$numObserved()+latentDigraph$numLatents(). The first latentDigraph$numObserved()
nodes correspond to the observed nodes in the graph, all other nodes are considered unobserved.
Throws an error if this is not true.
}
