% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{latentNodes}
\alias{latentNodes}
\alias{latentNodes.LatentDigraph}
\title{Get all latent nodes in the graph.}
\usage{
latentNodes(this, ...)

\method{latentNodes}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Get all latent nodes in the graph.
}
