% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{MixedGraph}
\alias{MixedGraph}
\title{Construct MixedGraph object}
\usage{
MixedGraph(L = matrix(0,1,1), O = matrix(0,1,1),
                  vertexNums = 1:nrow(L))
}
\arguments{
\item{L}{see \code{\link{graphID}} for the appropriate form of L.}

\item{O}{as for L.}

\item{vertexNums}{the labeling of the vertices in the graph in the order
of the rows of L and O. Labels must be positive integers.}
}
\value{
An object representing the MixedGraph
}
\description{
Creates an object representing a mixed graph. The methods that are currently
available to be used on the mixed graph include
\enumerate{
\item ancestors
\item descendants
\item parents
\item siblings
\item isSibling
\item htrFrom
\item trFrom
\item getHalfTrekSystem
\item getTrekSystem
\item inducedSubgraph
\item L
\item O
\item nodes
\item numNodes
\item stronglyConnectedComponent
\item tianComponent
\item tianDecompose
}
see the individual function documentation for more information.
}
