% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{MixedGraphFixedOrder}
\alias{MixedGraphFixedOrder}
\title{Construct MixedGraphFixedOrder object}
\usage{
MixedGraphFixedOrder(L = matrix(0,1,1), O = matrix(0,1,1))
}
\arguments{
\item{L}{see \code{\link{graphID}} for the appropriate form of L.}

\item{O}{as for L.}
}
\value{
An object representing the MixedGraphFixedOrder
}
\description{
Creates an object representing a mixed graph.
}
