% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mixedGraphHasSimpleNumbering}
\alias{mixedGraphHasSimpleNumbering}
\title{Checks that a MixedGraph has appropriate node numbering}
\usage{
mixedGraphHasSimpleNumbering(mixedGraph)
}
\arguments{
\item{mixedGraph}{the mixed graph object}
}
\description{
Checks that the input mixed graph has vertices are numbered from 1
to mixedGraph$numNodes(). Throws an error if they are not.
}
