% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{nodes}
\alias{nodes}
\alias{nodes.MixedGraph}
\title{Get all nodes in the graph.}
\usage{
nodes(this, ...)

\method{nodes}{MixedGraph}(this, ...)
}
\arguments{
\item{this}{the mixed graph object}

\item{...}{ignored.}
}
\description{
Get all nodes in the graph.
}
