% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_rules.R
\name{ntheta_rule}
\alias{ntheta_rule}
\title{N_theta rule (compares parameters to observed statistics)}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
N_theta rule (compares parameters to observed statistics)
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
