% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{null_byy_rule}
\alias{null_byy_rule}
\title{Null B_YY Rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Null B_YY Rule
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
