% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{numLatents}
\alias{numLatents}
\alias{numLatents.LatentDigraphFixedOrder}
\alias{numLatents.LatentDigraph}
\title{Number of latent nodes in the graph.}
\usage{
numLatents(this, ...)

\method{numLatents}{LatentDigraphFixedOrder}(this, ...)

\method{numLatents}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Number of latent nodes in the graph.
}
