% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{numNodes}
\alias{numNodes}
\alias{numNodes.MixedGraphFixedOrder}
\alias{numNodes.MixedGraph}
\title{Number of nodes in the graph.}
\usage{
numNodes(this)

\method{numNodes}{MixedGraphFixedOrder}(this)

\method{numNodes}{MixedGraph}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Number of nodes in the graph.
}
