% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{O}
\alias{O}
\alias{O.MixedGraph}
\title{Get adjacency matrix for bidirected part.}
\usage{
O(this)

\method{O}{MixedGraph}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Get adjacency matrix for bidirected part.
}
