% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{observedNodes}
\alias{observedNodes}
\alias{observedNodes.LatentDigraph}
\title{Get all observed nodes in the graph.}
\usage{
observedNodes(this, ...)

\method{observedNodes}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Get all observed nodes in the graph.
}
