% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{parents}
\alias{parents}
\alias{parents.MixedGraphFixedOrder}
\alias{parents.MixedGraph}
\title{All parents a collection of nodes.}
\usage{
parents(this, nodes)

\method{parents}{MixedGraphFixedOrder}(this, nodes)

\method{parents}{MixedGraph}(this, nodes)
}
\arguments{
\item{this}{the mixed graph object.}

\item{nodes}{nodes the nodes of which to find the parents.}
}
\value{
a vector of parents of the nodes.
}
\description{
All parents a collection of nodes.
}
