% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{plot.MixedGraph}
\alias{plot.MixedGraph}
\title{Plots the mixed graph}
\usage{
\method{plot}{MixedGraph}(x, ...)
}
\arguments{
\item{x}{the mixed graph object}

\item{...}{additional plotting arguments. Currently ignored.}
}
\description{
Plots the mixed graph
}
