% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_extmID.R
\name{print.extmIDresult}
\alias{print.extmIDresult}
\title{Prints a extmIDresult object}
\usage{
\method{print}{extmIDresult}(x, ...)
}
\arguments{
\item{x}{the extmIDresult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a extmIDresult object as returned by
\code{\link{extmID}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
