% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.GenericIDResult}
\alias{print.GenericIDResult}
\title{Prints a GenericIDResult object}
\usage{
\method{print}{GenericIDResult}(x, ...)
}
\arguments{
\item{x}{the GenericIDResult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a GenericIDResult object as returned by
\code{\link{generalGenericID}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
