% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rules.R
\name{print.lavid}
\alias{print.lavid}
\title{Printing function for rules list}
\usage{
\method{print}{lavid}(
  x,
  ...,
  names = c("", "Pass", "Necessary", "Sufficient"),
  include.msgs = TRUE,
  msgs.name = "Message",
  msgs.sec = "Messages",
  window = 56L,
  pos.lab = "Yes",
  neg.lab = "No",
  na.lab = "-"
)
}
\arguments{
\item{...}{Not currently used}

\item{names}{Character vector. The names of the columns of the main table}

\item{window}{Integer. The width of the output window}

\item{pos.lab}{Character. The label for positive cells, e.g., "Yes".}

\item{neg.lab}{Character. The label for negative cells, e.g., "No".}

\item{na.lab}{Character. The label for NA/blank cells.}

\item{rules}{A rules list produced by internal identification rules}

\item{warn}{Logical. If \code{TRUE} warnings are printed}

\item{warn.name}{Character. The name of the warning index column}

\item{warn.sec}{Character. The name of the warnings section}
}
\description{
Printing function for rules list
}
