% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfhtcID.R
\name{print.LfhtcIDResult}
\alias{print.LfhtcIDResult}
\title{Prints a LfhtcIDResult object}
\usage{
\method{print}{LfhtcIDResult}(x, ...)
}
\arguments{
\item{x}{the LfhtcIDResult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a LfhtcIDResult object as returned by
\code{\link{lfhtcID}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
