% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_mID.R
\name{print.mIDresult}
\alias{print.mIDresult}
\title{Prints a mIDresult object}
\usage{
\method{print}{mIDresult}(x, ...)
}
\arguments{
\item{x}{the mIDresult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a mIDresult object as returned by
\code{\link{mID}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
