% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.semid}
\alias{print.semid}
\title{Printing function for rules list}
\usage{
\method{print}{semid}(
  x,
  ...,
  names = c("", "Pass", "Necessary", "Sufficient"),
  include.msgs = TRUE,
  msgs.name = "Message",
  msgs.sec = "Messages",
  window = 56L,
  pos.lab = "Yes",
  neg.lab = "No",
  na.lab = "-"
)
}
\arguments{
\item{x}{A \code{semid} object, i.e., a list of rule output objects.}

\item{...}{Not currently used.}

\item{names}{Character vector. The names of the columns of the main table.}

\item{include.msgs}{Logical. If \code{TRUE} messages are printed.}

\item{msgs.name}{Character. The name of the messages index column.}

\item{msgs.sec}{Character. The name of the messages section.}

\item{window}{Integer. The width of the output window.}

\item{pos.lab}{Character. The label for positive cells, e.g., "Yes".}

\item{neg.lab}{Character. The label for negative cells, e.g., "No".}

\item{na.lab}{Character. The label for NA/blank cells.}
}
\description{
Printing function for rules list
}
