% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.SEMIDResult}
\alias{print.SEMIDResult}
\title{Prints a SEMIDResult object}
\usage{
\method{print}{SEMIDResult}(x, ...)
}
\arguments{
\item{x}{the SEMIDResult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a SEMIDResult object as returned by
\code{\link{semID}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
