% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.semscale}
\alias{print.semscale}
\title{Printing function for scaling table}
\usage{
\method{print}{semscale}(x, ..., include.msgs = TRUE, window = 56L)
}
\arguments{
\item{x}{A \code{semscale} object, i.e., a list of scaling information for each latent variable in the model.}

\item{...}{Not currently used.}

\item{include.msgs}{Logical. If \code{TRUE} messages are printed.}

\item{window}{Integer. The width of the output window.}
}
\description{
Printing function for scaling table
}
