% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_ZUTA.R
\name{print.ZUTAresult}
\alias{print.ZUTAresult}
\title{Prints a ZUTAresult object}
\usage{
\method{print}{ZUTAresult}(x, ...)
}
\arguments{
\item{x}{the ZUTAresult object}

\item{...}{optional parameters, currently unused.}
}
\description{
Prints a ZUTAresult object as returned by
\code{\link{ZUTA}}. Invisibly returns its argument via
\code{\link{invisible}(x)} as most print functions do.
}
