% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rules.R
\name{print_rules}
\alias{print_rules}
\title{Printing function for rules list}
\usage{
print_rules(
  rules,
  names = c("", "Pass", "Necessary", "Sufficient"),
  warn = TRUE,
  warn.name = "Warning",
  warn.sec = "Warnings",
  window = 56L,
  pos.lab = "Yes",
  neg.lab = "No",
  na.lab = "-"
)
}
\arguments{
\item{rules}{A rules list produced by internal identification rules}

\item{names}{Character vector. The names of the columns of the main table}

\item{warn}{Logical. If \code{TRUE} warnings are printed}

\item{warn.sec}{Character. The name of the warnings section}

\item{window}{Integer. The width of the output window}

\item{pos.lab}{Character. The label for positive cells, e.g., "Yes".}

\item{neg.lab}{Character. The label for negative cells, e.g., "No".}

\item{na.lab}{Character. The label for NA/blank cells.}

\item{warn.col}{Character. The name of the warning index column}
}
\description{
Printing function for rules list
}
\author{
Zach Vig
}
\keyword{internal}
