% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{recursive_corr_err_rule}
\alias{recursive_corr_err_rule}
\title{Recursive model with correlated errors rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Recursive model with correlated errors rule
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Brito, C., & Pearl, J. (2002). A new identification condition
for recursive models with correlated errors.
}
\author{
Zach Vig
}
\keyword{internal}
