% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{reg_rules}
\alias{reg_rules}
\alias{rule_reg_null_byy}
\alias{rule_reg_fully_recursive}
\alias{rule_reg_recursive_corr_err}
\title{Rules for simultaneous equation/regression models}
\usage{
rule_reg_null_byy(partable)

rule_reg_fully_recursive(partable)

rule_reg_recursive_corr_err(partable)
}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Rules for simultaneous equation/regression models
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).

Brito, C., & Pearl, J. (2002). A new identification condition
for recursive models with correlated errors.
}
\keyword{internal}
