% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scaling}
\alias{scaling}
\title{Check if latent variables are correctly scaled}
\usage{
scaling(partable, lv = NULL)
}
\arguments{
\item{partable}{A \code{lavaan} parameter table}

\item{lv}{An optional character vector of the specific latent variables you
would like to check. Otherwise, all latent variables are extracted from the
parameter table}
}
\value{
An (invisible) logical vector specifying whether each latent variable
(included as the names of the vector) is correctly scaled
}
\description{
This is a helper function for checking whether all latent variables in a model
are scaled.
}
\examples{
\dontrun{
mod1 <- ' l1 =~ x1 + x2
          l2 =~ x3 + x4 + x5 '
mod1.partable <- lavaan::lavaanify(mod1, auto = TRUE, model.type = "sem")
mod1.scaled <- scaling(mod1.partable, lv = "l2")
}

}
