% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semID-package.R
\docType{package}
\name{semID-package}
\alias{semID}
\alias{semID-package}
\title{semID package outline}
\description{
In order for structural equation model (SEM) estimation to be reliable, the model must be identified, i.e., every parameter in the model must have a unique solution. The SEM literature has a plethora of rules to help check if a model is identified, many of which are implemented in this package.
}
\section{Purpose}{

semID allows the user to input a Structural Equation Model (SEM) in [`lavaan`](https://lavaan.ugent.be/)
(Roseel, 2012) syntax and check it against a number of identification rules from the literature. Rules
are specified as being necessary and/or sufficient and specific reasons are given when a rule is broken.
Caution is given to the user in using the output of the package as the sole determinant of model identification
-- instead, it should be used as a "quick check" for any outstanding issues with the model.
}

\section{Main functions}{

The `id()` function is the work horse function of the package which evaluates a user-supplied model against
a number of identification rules. Calling the function in-line on a model string will print a table to the console
with the results of the rule checks. Assigning the output of `id()` to an object will save the results and can be
printed later.

The `scaling()` function lets the user verify if each latent variable is scaled in a model with latent variables.
The output includes whether the variable is scaled, the scaling indicator (if applicable), the method used to scale
it (if applicable), and the reason why it is not scaled (if applicable).

The `id2()` function uses the two-step rule for SEM identification in which (a) all structural paths are converted to
covariances, making the model a confirmatory factor analysis (CFA) model, (b) the CFA model is checked for identification,
(c) if the CFA is identified, the original SEM is converted into a simultaneous equations model (SEM) where all latent
variables are treated as observed, and (d) the resulting model is checked for identification. If both the CFA and SEM are
identified, the original SEM is identified. The `id2()` function prints the results of both steps to the console.
}

\section{Development notes}{

The core of the package was developed by Zach Vig, based on Ken Bollen's 'Elements of Structural Equation Models (SEMs)' (2026),
with style inspiration from the `lavaan` package. Additionally, OpenAI's GPT 5.2 and 5.4 mini models were used to (a) edit functions
and documentation for clarity, (b) implement the Depth-First Search algorithm for checking recursion, (c) help fix bugs, (d) help
write the test suite.
}

\author{
\strong{Maintainer}: Zach Vig \email{zachvig@unc.edu}

}
\keyword{internal}
