% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{stronglyConnectedComponent}
\alias{stronglyConnectedComponent}
\alias{stronglyConnectedComponent.LatentDigraphFixedOrder}
\alias{stronglyConnectedComponent.LatentDigraph}
\alias{stronglyConnectedComponent.MixedGraph}
\title{Strongly connected component}
\usage{
stronglyConnectedComponent(this, node, ...)

\method{stronglyConnectedComponent}{LatentDigraphFixedOrder}(this, node, ...)

\method{stronglyConnectedComponent}{LatentDigraph}(this, node, ...)

\method{stronglyConnectedComponent}{MixedGraph}(this, node, ...)
}
\arguments{
\item{this}{the graph object}

\item{node}{the node for which to get the strongly connected component.}

\item{...}{ignored.}
}
\description{
Get the strongly connected component for a node i in the graph the graph.
}
