% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{stronglyConnectedComponents}
\alias{stronglyConnectedComponents}
\alias{stronglyConnectedComponents.MixedGraphFixedOrder}
\title{Strongly connected components}
\usage{
stronglyConnectedComponents(this)

\method{stronglyConnectedComponents}{MixedGraphFixedOrder}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Get the strongly connected components of a graph
}
