% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetsOfSize}
\alias{subsetsOfSize}
\title{Returns all subsets of a certain size}
\usage{
subsetsOfSize(x, k)
}
\arguments{
\item{x}{a vector from which to get subsets}

\item{k}{the size of the subsets returned}
}
\value{
a list of all subsets of x of a given size k
}
\description{
For an input vector x, returns in a list, the collection of all subsets
of x of size k.
}
