% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{sur_rule}
\alias{sur_rule}
\title{Seemingly Unrelated Regressions rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Seemingly Unrelated Regressions rule
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
