% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_rules.R
\name{three_indicator_rule}
\alias{three_indicator_rule}
\title{Three Indicator rules}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Three Indicator rules
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).
}
\author{
Zach Vig
}
\keyword{internal}
