% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{tianComponent}
\alias{tianComponent}
\alias{tianComponent.MixedGraph}
\title{Returns the Tian c-component of a node}
\usage{
tianComponent(this, node)

\method{tianComponent}{MixedGraph}(this, node)
}
\arguments{
\item{this}{the mixed graph object}

\item{node}{the node for which to return its c-component}
}
\description{
Returns the Tian c-component of a node
}
