% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{tianDecompose}
\alias{tianDecompose}
\alias{tianDecompose.MixedGraph}
\title{Performs the tian decomposition on the mixed graph}
\usage{
tianDecompose(this)

\method{tianDecompose}{MixedGraph}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Uses the Tian decomposition to break the mixed graph into c-components.
These c-components are slightly different than those from Tian (2005)
in that if they graph is not acyclic the bidirected components are
combined whenever they are connected by a directed loop.
}
\references{
Jin Tian. 2005. Identifying direct causal effects in linear models. In
\emph{Proceedings of the 20th national conference on Artificial intelligence
- Volume 1} (AAAI'05), Anthony Cohn (Ed.), Vol. 1. AAAI Press 346-352.
}
