% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tianIdentifier}
\alias{tianIdentifier}
\title{Identifies components in a tian decomposition}
\usage{
tianIdentifier(idFuncs, cComponents)
}
\arguments{
\item{idFuncs}{a list of identifier functions for the c-components}

\item{cComponents}{the c-components of the mixed graph as returned by
\code{\link{tianDecompose}}.}
}
\value{
a new identifier function
}
\description{
Creates an identification function which combines the identification
functions created on a collection of c-components into a identification
for the full mixed graph.
}
