% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{toEx}
\alias{toEx}
\alias{toEx.MixedGraph}
\title{Transforms a vector of node indices in the internal rep. into external numbering}
\usage{
toEx(this, nodes)

\method{toEx}{MixedGraph}(this, nodes)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{the nodes to transform}
}
\description{
Transforms a vector of node indices in the internal rep. into external numbering
}
