% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{toIn}
\alias{toIn}
\alias{toIn.MixedGraph}
\title{Transforms a vector of given node indices into their internal numbering}
\usage{
toIn(this, nodes)

\method{toIn}{MixedGraph}(this, nodes)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{the nodes to transform}
}
\description{
Transforms a vector of given node indices into their internal numbering
}
