% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{trFrom}
\alias{trFrom}
\alias{trFrom.MixedGraphFixedOrder}
\alias{trFrom.MixedGraph}
\title{Trek reachable nodes.}
\usage{
trFrom(this, node)

\method{trFrom}{MixedGraphFixedOrder}(this, node)

\method{trFrom}{MixedGraph}(this, node)
}
\arguments{
\item{this}{the mixed graph object}

\item{node}{the node from which to find trek-reachable nodes.}
}
\description{
Like \code{\link{htrFrom}} but for the nodes that are trek-reachable from a
node
}
