% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_rules.R
\name{two_emitted_paths_rule}
\alias{two_emitted_paths_rule}
\title{2+ Emitted Paths rule}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
2+ Emitted Paths rule
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen & Davis (2009). Two Rules of Identification for Structural Equation Models.
}
\author{
Zach Vig
}
\keyword{internal}
