% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowGraph.R
\name{updateVertexCapacities}
\alias{updateVertexCapacities}
\alias{updateVertexCapacities.FlowGraph}
\title{Update vertex capacities.}
\usage{
updateVertexCapacities(this, vertices, newCaps)

\method{updateVertexCapacities}{FlowGraph}(this, vertices, newCaps)
}
\arguments{
\item{this}{the flow graph object}

\item{vertices}{the vertices to update.}

\item{newCaps}{the new capacities for the vertices.}
}
\description{
Update vertex capacities.
}
