% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfhtcID.R
\name{validateLatentNodesAreSources}
\alias{validateLatentNodesAreSources}
\title{A helper function to validate that latent nodes in a LatentDigraph are sources.}
\usage{
validateLatentNodesAreSources(graph)
}
\arguments{
\item{graph}{the LatentDigraph}
}
\description{
Produces an error if not all latent nodes are sources.
}
