% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateMatrix}
\alias{validateMatrix}
\title{A helper function to validate an input matrix.}
\usage{
validateMatrix(L)
}
\arguments{
\item{L}{See above description.}
}
\value{
No return value
}
\description{
This helper function validates that an input matrix, L, is of the
the appropriate form to be interpreted by the other functions. In particular
it should be square matrix of 1's and 0's with all 0's along its
diagonal. If any of the above conditions is not met, this function will
throw an error.
}
