% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{validateNodes}
\alias{validateNodes}
\title{A helper function to validate if input nodes are valid.}
\usage{
validateNodes(nodes, numNodes)
}
\arguments{
\item{nodes}{the input nodes, expected to be from the collection
1:(number of nodes in the graph)}

\item{numNodes}{the number of observed nodes in the graph.}
}
\description{
Produces an error if outside bounds.
}
