% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{validateVarArgsEmpty}
\alias{validateVarArgsEmpty}
\title{A helper function to validate that there are no variable arguments}
\usage{
validateVarArgsEmpty(...)
}
\arguments{
\item{...}{the variable arguments}
}
\description{
Produces an error if there are variable arguments.
}
