% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_ZUTA.R
\name{ZUTA}
\alias{ZUTA}
\title{Check the Zero Upper Triangular Assumption}
\usage{
ZUTA(lambda)
}
\arguments{
\item{lambda}{adjacency matrix with number of cols = number of latent nodes,
number of rows = number of observed nodes}
}
\value{
a Boolean, whether the graph fulfills ZUTA
}
\description{
Check the Zero Upper Triangular Assumption
}
\references{
Sturma, N., Kranzlmüller, M., Portakal, I., and Drton, M.  (2025) Matching
Criterion for Identifiability in Sparse Factor Analysis.
arXiv:2502.02986
}
