% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{ancestors}
\alias{ancestors}
\alias{ancestors.MixedGraphFixedOrder}
\alias{ancestors.MixedGraph}
\title{All ancestors of a collection of nodes}
\usage{
ancestors(this, nodes)

\method{ancestors}{MixedGraphFixedOrder}(this, nodes)

\method{ancestors}{MixedGraph}(this, nodes)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{the nodes from which to find all ancestors}
}
\description{
Finds all the ancestors of a collection of nodes. These ancestors DO include
the nodes themselves (every node is considered an ancestor of itself).
}
