% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{ancestors}
\alias{ancestors}
\title{Get ancestors of nodes in a graph.}
\usage{
ancestors(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the ancestors.}
}
\value{
a sorted vector of all ancestor nodes.
}
\description{
Get the ancestors of a collection of nodes in a graph g, the ancestors DO
include the the nodes themselves.
}

