% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{check_recursion}
\alias{check_recursion}
\title{Function to check if a simultaneous equations model is recursive}
\usage{
check_recursion(partable, base, comp)
}
\description{
Function to check if a simultaneous equations model is recursive
}
\keyword{internal}
