% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_localBBCriterion.R
\name{checkLocalBBCriterion}
\alias{checkLocalBBCriterion}
\title{Check Local BB-Criterion}
\usage{
checkLocalBBCriterion(adjMatrix, latentNodes, observedNodes)
}
\arguments{
\item{adjMatrix}{of the graph}

\item{latentNodes}{set of latent nodes}

\item{observedNodes}{set of observed nodes}
}
\value{
a list consisting of a Boolean, whether the graph is
        sign-identifiable and if yes, a list consisting of the sets
}
\description{
checks whether a graph is recursively identifiable by only using the local
BB-criterion
}
\references{
Sturma, N., Kranzlmüller, M., Portakal, I., and Drton, M.  (2025) Matching
Criterion for Identifiability in Sparse Factor Analysis.
arXiv:2502.02986
}
