% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{createTrekFlowGraph}
\alias{createTrekFlowGraph}
\alias{createTrekFlowGraph.MixedGraphFixedOrder}
\title{Helper function to create a flow graph.}
\usage{
createTrekFlowGraph(this)

\method{createTrekFlowGraph}{MixedGraphFixedOrder}(this)
}
\arguments{
\item{this}{the mixed graph object}
}
\description{
Helper function to create a flow graph.
}
