% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{createTrGraph}
\alias{createTrGraph}
\alias{createTrGraph.LatentDigraphFixedOrder}
\title{Helper function to create a graph encoding trek reachable relationships.}
\usage{
createTrGraph(this, ...)

\method{createTrGraph}{LatentDigraphFixedOrder}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Helper function to create a graph encoding trek reachable relationships.
}
