% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{descendants}
\alias{descendants}
\alias{descendants.LatentDigraphFixedOrder}
\alias{descendants.LatentDigraph}
\alias{descendants.MixedGraph}
\title{Get descendants of a collection of observed nodes}
\usage{
descendants(this, nodes, ...)

\method{descendants}{LatentDigraphFixedOrder}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{descendants}{LatentDigraph}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{descendants}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object}

\item{nodes}{the nodes from which to get the descendants.}

\item{...}{ignored.}

\item{includeObserved}{if TRUE includes observed nodes in the returned set.}

\item{includeLatents}{if TRUE includes latent nodes in the returned set.}
}
\description{
Finds all descendants of a collection of nodes, this DOES include the nodes
themselves (every node is considered a descendant of itself).
}
