% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{descendants}
\alias{descendants}
\title{Getdescendants of nodes in a graph.}
\usage{
descendants(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the descendants.}
}
\value{
a sorted vector of all descendants of nodes.
}
\description{
Gets the descendants of a collection of nodes in a graph (all nodes that can
be reached by following directed edges from those nodes). Descendants DO
include the nodes themselves.
}

