% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{get_rules}
\alias{get_rules}
\title{Gather identification rules as a list}
\usage{
get_rules(rule = "*", model_type = "*")
}
\arguments{
\item{rule}{A string specifying the name of the rule as it's defined
in the package. Use "*" to get all rules (or all rules of the defined
model type)}

\item{model_type}{A string specifying the model sub-type from which to get
rules. Sub-types include "reg" (simultaneous equation models/regression models) and
"cfa" (confirmatory factor analysis models). Use "sem" to get rules that apply
to all structural equation models. Use "*" to get all rules in the package.}
}
\value{
A list object with the rule function(s) specified by the user.
}
\description{
The `semID` package stores identification rule functions internally. This function
 makes them available to the user as a list.
}
\examples{
\dontrun{
  rules <- get_rules(rule = "*", model_type = "cfa")
}

}
\author{
Zach Vig
}
