% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{getMixedGraph}
\alias{getMixedGraph}
\alias{getMixedGraph.LatentDigraph}
\title{Get the corresponding mixed graph}
\usage{
getMixedGraph(this, ...)

\method{getMixedGraph}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the LatentDigraph object}

\item{...}{ignored}
}
\description{
Only works for graphs where the latent nodes are source nodes
}
