% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{getTrekSystem}
\alias{getTrekSystem}
\alias{getTrekSystem.MixedGraphFixedOrder}
\alias{getTrekSystem.MixedGraph}
\title{Determines if a trek system exists in the mixed graph.}
\usage{
getTrekSystem(this, fromNodes, toNodes, avoidEdgesOnRight)

\method{getTrekSystem}{MixedGraphFixedOrder}(this, fromNodes, toNodes,
  avoidEdgesOnRight = NULL)

\method{getTrekSystem}{MixedGraph}(this, fromNodes, toNodes,
  avoidEdgesOnRight = NULL)
}
\arguments{
\item{this}{the mixed graph object}

\item{fromNodes}{the start nodes}

\item{toNodes}{the end nodes}

\item{avoidEdgesOnRight}{a collection of edges in the graph that should not
be used on any right hand side of any trek in the
trek system.}
}
\description{
Determines if a trek system exists in the mixed graph.
}
