% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{graphID.decompose}
\alias{graphID.decompose}
\title{Determine generic identifiability by Tian Decomposition and HTC}
\usage{
graphID.decompose(L, O, decomp.if.acyclic = TRUE, test.globalID = TRUE,
  test.genericID = TRUE, test.nonID = TRUE)
}
\arguments{
\item{L}{Adjacency matrix for the directed part of the path
diagram/mixed graph; an edge pointing from i to j is encoded as L[i,j]=1.}

\item{O}{Adjacency matrix for the bidirected part of the path
diagram/mixed graph.}

\item{decomp.if.acyclic}{A logical value indicating whether an input graph
that is acyclic is to be decomposed before applying identifiability criteria.}

\item{test.globalID}{A logical value indicating whether or not global
identifiability is checked.}

\item{test.genericID}{A logical value indicating whether or not a sufficient
condition for generic identifiability is checked.}

\item{test.nonID}{A logical value indicating whether or not a condition
implying generic non-identifiability is checked.}
}
\value{
A list with two named components:

  1. Components - a list of lists. Each list represents one mixed Tian component
      of the graph. Each list contains named components corresponding to which
      nodes are in the component and results of various tests of
      identifiability on the component (see the parameter descriptions).

 2. Decomp - true if a decomposition occured, false if not.
}
\description{
Split a graph into mixed Tiancomponents and solve each separately
using the HTC.
}

