% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{htrFrom}
\alias{htrFrom}
\alias{htrFrom.MixedGraph}
\title{Half trek reachable nodes.}
\usage{
htrFrom(this, nodes, ...)

\method{htrFrom}{MixedGraph}(
  this,
  nodes,
  avoidLeftNodes = integer(0),
  avoidRightNodes = integer(0),
  ...
)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{the nodes from which to get all half-trek reachable nodes.}

\item{...}{ignored.}

\item{avoidLeftNodes}{a collection of nodes to avoid on the left}

\item{avoidRightNodes}{a collection of nodes to avoid on the right}
}
\value{
a vector of all nodes half-trek reachable from node.
}
\description{
Half trek reachable nodes.
}
