% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{Evaluate common Structural Equation Model (SEM) identification rules}
\usage{
id(model = NULL, warn = TRUE, call = "sem", ...)
}
\arguments{
\item{model}{A character string model in \code{lavaan} syntax, a
\code{lavaan} parameter table, or a fitted \code{lavaan} object.}

\item{warn}{Logical. If \code{TRUE}, the output will include why a rule does
not pass or is not applicable, along with any other helpful information.}

\item{call}{A character string specifying the call you intend to use to fit
the model. This will ensure the correct model defaults are specified. Options
currently include "lavaan", "sem", or "cfa". If a parameter table for fit
object are supplied, this argument is ignored.}

\item{...}{Additional arguments passed to the \code{lavaanify} function from
\code{lavaan}. See \link[lavaan]{lavaanify} for more information.}
}
\value{
Nothing. Returns once summary table is printed.
}
\description{
...
}
\examples{
# The Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
id(HS.model, warn = TRUE, call = "cfa", meanstructure = FALSE)
}
