% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{Evaluate common Structural Equation Model (SEM) identification rules.}
\usage{
id(model = NULL, warn = TRUE, ...)
}
\arguments{
\item{model}{A character string model in \code{lavaan} syntax or a
\code{lavaan} parameter table.}

\item{warn}{Logical. If \code{TRUE}, the output will include why a rule does
not pass.}

\item{...}{Additional arguments passed to the \code{lavaanify} function from
\code{lavaan}. See \link[lavaan]{lavaanify} for more information.}
}
\value{
Nothing. Returns once summary table is printed.
}
\description{
...
}
\examples{
# The Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
id(HS.model, warn = TRUE, meanstructure = FALSE)
}
