% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{inducedSubgraph}
\alias{inducedSubgraph}
\alias{inducedSubgraph.MixedGraph}
\title{Get the induced subgraph on a collection of nodes}
\usage{
inducedSubgraph(this, nodes)

\method{inducedSubgraph}{MixedGraph}(this, nodes)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{the nodes on which to create the induced subgraph.}
}
\description{
Get the induced subgraph on a collection of nodes
}
