% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{L}
\alias{L}
\alias{L.LatentDigraphFixedOrder}
\alias{L.LatentDigraph}
\alias{L.MixedGraph}
\title{Get directed adjacency matrix.}
\usage{
L(this, ...)

\method{L}{LatentDigraphFixedOrder}(this, ...)

\method{L}{LatentDigraph}(this, ...)

\method{L}{MixedGraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored.}
}
\description{
Get directed adjacency matrix.
}
