% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_mID.R
\name{mID}
\alias{mID}
\title{Check M-Identifiability.}
\usage{
mID(lambda, maxCard = length(observedNodes))
}
\arguments{
\item{lambda}{adjacency matrix with number of cols = number of latent nodes,
number of rows = number of observed nodes}

\item{maxCard}{(optional) maximum size of set W}
}
\value{
a list consisting of a Boolean, whether the graph is
        sign-identifiable and if yes, a list consisting of the sets
}
\description{
Check M-Identifiability.
}
\references{
Sturma, N., Kranzlmüller, M., Portakal, I., and Drton, M.  (2025) Matching
Criterion for Identifiability in Sparse Factor Analysis.
arXiv:2502.02986
}
