% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{numNodes}
\alias{numNodes}
\alias{numNodes.LatentDigraphFixedOrder}
\alias{numNodes.LatentDigraph}
\alias{numNodes.MixedGraph}
\title{Number of nodes in the graph.}
\usage{
numNodes(this, ...)

\method{numNodes}{LatentDigraphFixedOrder}(this, ...)

\method{numNodes}{LatentDigraph}(this, ...)

\method{numNodes}{MixedGraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored.}
}
\description{
Number of nodes in the graph.
}
