% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{numObserved}
\alias{numObserved}
\alias{numObserved.LatentDigraphFixedOrder}
\alias{numObserved.LatentDigraph}
\title{Number of observed nodes in the graph.}
\usage{
numObserved(this, ...)

\method{numObserved}{LatentDigraphFixedOrder}(this, ...)

\method{numObserved}{LatentDigraph}(this, ...)
}
\arguments{
\item{this}{the graph object}

\item{...}{ignored}
}
\description{
Number of observed nodes in the graph.
}
