% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R
\name{observedParents}
\alias{observedParents}
\alias{observedParents.LatentDigraph}
\title{Get the observed parents on a collection of nodes}
\usage{
observedParents(this, nodes, ...)

\method{observedParents}{LatentDigraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object}

\item{nodes}{the nodes on which to get the observed parents}

\item{...}{ignored}
}
\description{
Get the observed parents on a collection of nodes
}
