% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{parents}
\alias{parents}
\title{Get parents of nodes in a graph.}
\usage{
parents(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the parents.}
}
\value{
a sorted vector of all parent nodes.
}
\description{
Get the parents of a collection of nodes in a graph g, the parents DO include
the input nodes themselves.
}

