% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{parents}
\alias{parents}
\alias{parents.LatentDigraphFixedOrder}
\alias{parents.LatentDigraph}
\alias{parents.MixedGraph}
\title{All parents of a collection of nodes.}
\usage{
parents(this, nodes, ...)

\method{parents}{LatentDigraphFixedOrder}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{parents}{LatentDigraph}(this, nodes, includeObserved = T, includeLatents = T, ...)

\method{parents}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the graph object.}

\item{nodes}{nodes the nodes of which to find the parents.}

\item{...}{ignored.}

\item{includeObserved}{if TRUE includes observed nodes in the returned set.}

\item{includeLatents}{if TRUE includes latent nodes in the returned set.}
}
\value{
the observed parents.
}
\description{
Returns all parents of the collection (does not necessarily include the
input nodes themselves unless they are parents of one another).
}
