% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{plot.LatentDigraph}
\alias{plot.LatentDigraph}
\alias{plot.MixedGraph}
\title{Plots the latent digraph}
\usage{
\method{plot}{LatentDigraph}(x, ...)

\method{plot}{MixedGraph}(x, ...)
}
\arguments{
\item{x}{the mixed graph object}

\item{...}{additional plotting arguments. Currently ignored.}
}
\description{
Plots the latent digraph

Plots the mixed graph
}
