% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.semid2}
\alias{print.semid2}
\title{Printing function for two-step rules lists}
\usage{
\method{print}{semid2}(
  x,
  ...,
  step.names = c("Measurement Model", "Latent Variable/Structural Model"),
  step.titles = c("STEP 1", "STEP 2"),
  div = strrep("-", 56L)
)
}
\arguments{
\item{x}{A \code{semid2} object}

\item{...}{Arguments to be passed to print.semid.}

\item{step.names}{Character. The names of each step/block.}

\item{step.titles}{Character. The labels to be given to each step.}

\item{div}{Character. The divider to be used before each step/block.}
}
\description{
Printing function for two-step rules lists
}
