% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rules.R
\name{reg_rules}
\alias{reg_rules}
\alias{rule_reg_null_byy}
\alias{rule_reg_fully_recursive}
\alias{rule_reg_recursive_corr_err}
\title{Rules for simultaneous equations/regression models}
\usage{
rule_reg_null_byy(partable)

rule_reg_fully_recursive(partable)

rule_reg_recursive_corr_err(partable)
}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Simultaneous equations models are structural models without any latent
variables. Regression models are a subset of these models with a single
outcome, while the general simultaneous equations model can accomodate
multiple outcomes
}
\details{
\itemize{
 \item{Null B_YY Rule}{No endogenous variable is the predictor of another
 endogenous variable. Sufficient but not necessary.}
 \item{Fully Recursive Model Rule}{The model has no feedback loops (i.e.,
 is recursive) and there are no correlated errors ("fully" recursive).
 Sufficient but not necessary.}
 \item{Recursive with Correlated Errors Rule}{The model has no feedback loops
 (i.e., is recursive) but can have correlated errors as long as the errors are
 not for terms between which exists a structural/directional path. Sufficient
 but not necessary.}
}
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Brito, C., & Pearl, J. (2002). A new identification condition
for recursive models with correlated errors.
}
\keyword{internal}
