% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_rules.R
\name{sem_rules}
\alias{sem_rules}
\alias{rule_sem_ntheta}
\alias{rule_sem_latent_scaling}
\alias{rule_sem_two_emitted_paths}
\alias{rule_sem_exogenous_x}
\title{Rules for all structural equation models}
\usage{
rule_sem_ntheta(partable)

rule_sem_latent_scaling(partable)

rule_sem_two_emitted_paths(partable)

rule_sem_exogenous_x(partable)
}
\arguments{
\item{partable}{A \code{lavaan} parameter table}
}
\description{
Rules for all structural equation models
}
\references{
Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen, K. A., Lilly, A. G., & Luo, L. (2024). Selecting scaling
indicators in structural equation models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen & Davis (2009). Two Rules of Identification for Structural Equation Models.

Bollen (2026). Elements of Structural Equation Models (SEMs).

Bollen & Davis (2009). Two Rules of Identification for Structural Equation Models.
}
\keyword{internal}
