% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMID.R
\docType{package}
\name{SEMID-package}
\alias{SEMID}
\alias{SEMID-package}
\title{Test package documentation.}
\description{
Test package documentation description.
}
\details{
The only function you're likely to need from \pkg{SEMID} is
\code{\link{graphID}}. Otherwise refer to the individual function
documentation.
}

