% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMID.R
\docType{package}
\name{SEMID-package}
\alias{SEMID}
\alias{SEMID-package}
\title{SEMID package documentation.}
\description{
SEMID provides a number of methods for testing the global/generic
identifiability of mixed graphs.
}
\details{
The only function you're likely to need from \pkg{SEMID} is
\code{\link{graphID}}. Otherwise refer to the individual function
documentation.
}

