% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{siblings}
\alias{siblings}
\alias{siblings.MixedGraph}
\title{All siblings of a collection of nodes}
\usage{
siblings(this, nodes, ...)

\method{siblings}{MixedGraph}(this, nodes, ...)
}
\arguments{
\item{this}{the mixed graph object}

\item{nodes}{a vector of nodes of which to find the siblings.}

\item{...}{ignored.}
}
\value{
a vector of all of the siblings.
}
\description{
All siblings of a collection of nodes
}
