% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphID.R
\name{siblings}
\alias{siblings}
\title{Get siblings of nodes in a graph.}
\usage{
siblings(g, nodes)
}
\arguments{
\item{g}{the graph (as an igraph).}

\item{nodes}{the nodes in the graph of which to get the siblings.}
}
\value{
a sorted vector of all siblings of nodes.
}
\description{
Get the siblings of a collection of nodes in a graph g, the siblings DO
include the input nodes themselves.
}

