% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGraph.R
\name{stronglyConnectedComponent}
\alias{stronglyConnectedComponent}
\alias{stronglyConnectedComponent.MixedGraphFixedOrder}
\alias{stronglyConnectedComponent.MixedGraph}
\title{Strongly connected component}
\usage{
stronglyConnectedComponent(this, node)

\method{stronglyConnectedComponent}{MixedGraphFixedOrder}(this, node)

\method{stronglyConnectedComponent}{MixedGraph}(this, node)
}
\arguments{
\item{this}{the mixed graph object}

\item{node}{the node for which to get the strongly connected component.}
}
\description{
Get the strongly connected component for a node i in the directed part of
the graph.
}
