% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDigraph.R, R/MixedGraph.R
\name{trFrom}
\alias{trFrom}
\alias{trFrom.LatentDigraphFixedOrder}
\alias{trFrom.LatentDigraph}
\alias{trFrom.MixedGraph}
\title{Trek reachable nodes.}
\usage{
trFrom(this, nodes, ...)

\method{trFrom}{LatentDigraphFixedOrder}(
  this,
  nodes,
  avoidLeftNodes = integer(0),
  avoidRightNodes = integer(0),
  includeObserved = T,
  includeLatents = T,
  ...
)

\method{trFrom}{LatentDigraph}(
  this,
  nodes,
  avoidLeftNodes = integer(0),
  avoidRightNodes = integer(0),
  includeObserved = T,
  includeLatents = T,
  ...
)

\method{trFrom}{MixedGraph}(
  this,
  nodes,
  avoidLeftNodes = integer(0),
  avoidRightNodes = integer(0),
  ...
)
}
\arguments{
\item{this}{the graph object}

\item{nodes}{the nodes from which to find trek-reachable nodes.}

\item{...}{ignored.}

\item{avoidLeftNodes}{a collection of nodes to avoid on the left}

\item{avoidRightNodes}{a collection of nodes to avoid on the right}

\item{includeObserved}{if TRUE includes observed nodes in the returned set.}

\item{includeLatents}{if TRUE includes latent nodes in the returned set.}
}
\description{
Gets all nodes that are trek reachable from a collection of nodes.
}
