% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowGraph.R
\name{updateEdgeCapacities}
\alias{updateEdgeCapacities}
\alias{updateEdgeCapacities.FlowGraph}
\title{Update edge capacities.}
\usage{
updateEdgeCapacities(this, edges, newCaps)

\method{updateEdgeCapacities}{FlowGraph}(this, edges, newCaps)
}
\arguments{
\item{this}{the flow graph object}

\item{edges}{the vertices to update (as a 2xr matrix with ith row
corresponding to the edge edges[i,1]->edges[i,2].}

\item{newCaps}{the new capacities for the edges}
}
\description{
Update edge capacities.
}
